more test programs

+++ sw/airborne/fms/fms_autopilot_msg.h
@@ -4,9 +4,9 @@
 #include <inttypes.h>
 
 struct AutopilotMessageFoo {
-  uint32_t foo;
-  uint32_t bar;
-  uint32_t blaa;
+  uint8_t foo;
+  uint8_t bar;
+  uint8_t blaa;
 };
